+++ examples/example-remote/src/pages/WorkflowCase/Edit/FormInfo/index.jsx
@@ -895,6 +895,7 @@ class BasicInfo extends TabPageBase {
             title: title || {},
             items,
           }}
+          formItems={listDataSchema}
           approveList={isArray(listApprove) ? listApprove : []}
           allApproveProcessList={listChainApproveAdjust}
           signetStyle={signetStyle}