+++ examples/example-remote/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.4.165
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.169
+  - antd-management-fast-design-react@1.0.104
+  - antd-management-fast-component@2.7.21
+  - antd-management-fast-framework@2.12.24
+  - antd-management-fast-formily@1.0.116
+  - antd-management-fast-common@2.6.134
+  - antd-management-fast-flow@1.0.104
+
 ## 1.4.164
 
 ### Patch Changes