+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,16 @@
 # antd-management-fast-design-playground
 
+## 1.0.169
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.104
+  - antd-management-fast-component@2.7.21
+  - antd-management-fast-formily@1.0.116
+  - antd-management-fast-common@2.6.134
+
 ## 1.0.168
 
 ### Patch Changes